Refactor font imports and add Cypress configuration
